# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * La Grand Pharmacie d'Alger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductPole = (typeof ProductPole)[keyof typeof ProductPole];

export const ProductPole = {
  medical: "medical",
  paramedical: "paramedical",
  cosmetique: "cosmetique",
  supplements: "supplements",
} as const;

export interface Product {
  id: number;
  name: string;
  description?: string;
  pole: ProductPole;
  category: string;
  besoin?: string;
  price: number;
  imageUrl?: string;
  ingredients?: string;
  posologie?: string;
  contreIndications?: string;
  texture?: string;
  origin?: string;
  isFeatured: boolean;
  inStock: boolean;
  badge?: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  offset: number;
  limit: number;
}

export type CategoryPole = (typeof CategoryPole)[keyof typeof CategoryPole];

export const CategoryPole = {
  medical: "medical",
  paramedical: "paramedical",
  cosmetique: "cosmetique",
  supplements: "supplements",
} as const;

export interface Category {
  id: number;
  name: string;
  slug: string;
  pole: CategoryPole;
  productCount: number;
}

export type CatalogSummaryPoleCounts = { [key: string]: number };

export interface Besoin {
  id: number;
  name: string;
  slug: string;
  pole?: string;
  productCount: number;
  icon?: string;
}

export interface CatalogSummary {
  totalProducts: number;
  poleCounts: CatalogSummaryPoleCounts;
  featuredBesoins: Besoin[];
  newArrivals: Product[];
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  productImageUrl?: string;
  pole: string;
  price: number;
  quantity: number;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartBody {
  sessionId: string;
  productId: number;
  quantity: number;
}

export interface UploadPrescriptionBody {
  patientName: string;
  phone: string;
  fileDataUrl: string;
  notes?: string;
}

export interface PrescriptionConfirmation {
  id: number;
  reference: string;
  status: string;
  message: string;
  estimatedReady?: string;
}

export type ListProductsParams = {
  pole?: ListProductsPole;
  category?: string;
  besoin?: string;
  search?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};

export type ListProductsPole =
  (typeof ListProductsPole)[keyof typeof ListProductsPole];

export const ListProductsPole = {
  medical: "medical",
  paramedical: "paramedical",
  cosmetique: "cosmetique",
  supplements: "supplements",
} as const;

export type ListCategoriesParams = {
  pole?: ListCategoriesPole;
};

export type ListCategoriesPole =
  (typeof ListCategoriesPole)[keyof typeof ListCategoriesPole];

export const ListCategoriesPole = {
  medical: "medical",
  paramedical: "paramedical",
  cosmetique: "cosmetique",
  supplements: "supplements",
} as const;

export type GetFeaturedProductsParams = {
  pole?: GetFeaturedProductsPole;
  limit?: number;
};

export type GetFeaturedProductsPole =
  (typeof GetFeaturedProductsPole)[keyof typeof GetFeaturedProductsPole];

export const GetFeaturedProductsPole = {
  medical: "medical",
  paramedical: "paramedical",
  cosmetique: "cosmetique",
  supplements: "supplements",
} as const;

export type GetCartParams = {
  sessionId: string;
};