# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * La Grand Pharmacie d'Alger API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List products with optional filters
 */
export const listProductsQueryLimitDefault = 24;
export const listProductsQueryOffsetDefault = 0;

export const ListProductsQueryParams = zod.object({
  pole: zod
    .enum(["medical", "paramedical", "cosmetique", "supplements"])
    .optional(),
  category: zod.coerce.string().optional(),
  besoin: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  featured: zod.coerce.boolean().optional(),
  limit: zod.coerce.number().default(listProductsQueryLimitDefault),
  offset: zod.coerce.number().default(listProductsQueryOffsetDefault),
});

export const ListProductsResponse = zod.object({
  products: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      description: zod.string().optional(),
      pole: zod.enum(["medical", "paramedical", "cosmetique", "supplements"]),
      category: zod.string(),
      besoin: zod.string().optional(),
      price: zod.number(),
      imageUrl: zod.string().optional(),
      ingredients: zod.string().optional(),
      posologie: zod.string().optional(),
      contreIndications: zod.string().optional(),
      texture: zod.string().optional(),
      origin: zod.string().optional(),
      isFeatured: zod.boolean(),
      inStock: zod.boolean(),
      badge: zod.string().optional(),
    }),
  ),
  total: zod.number(),
  offset: zod.number(),
  limit: zod.number(),
});

/**
 * @summary Get product by ID
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  pole: zod.enum(["medical", "paramedical", "cosmetique", "supplements"]),
  category: zod.string(),
  besoin: zod.string().optional(),
  price: zod.number(),
  imageUrl: zod.string().optional(),
  ingredients: zod.string().optional(),
  posologie: zod.string().optional(),
  contreIndications: zod.string().optional(),
  texture: zod.string().optional(),
  origin: zod.string().optional(),
  isFeatured: zod.boolean(),
  inStock: zod.boolean(),
  badge: zod.string().optional(),
});

/**
 * @summary List categories, optionally filtered by pole
 */
export const ListCategoriesQueryParams = zod.object({
  pole: zod
    .enum(["medical", "paramedical", "cosmetique", "supplements"])
    .optional(),
});

export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  pole: zod.enum(["medical", "paramedical", "cosmetique", "supplements"]),
  productCount: zod.number(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Homepage summary — product counts per pole, featured products, besoins
 */
export const GetCatalogSummaryResponse = zod.object({
  totalProducts: zod.number(),
  poleCounts: zod.record(zod.string(), zod.number()),
  featuredBesoins: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      slug: zod.string(),
      pole: zod.string().optional(),
      productCount: zod.number(),
      icon: zod.string().optional(),
    }),
  ),
  newArrivals: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      description: zod.string().optional(),
      pole: zod.enum(["medical", "paramedical", "cosmetique", "supplements"]),
      category: zod.string(),
      besoin: zod.string().optional(),
      price: zod.number(),
      imageUrl: zod.string().optional(),
      ingredients: zod.string().optional(),
      posologie: zod.string().optional(),
      contreIndications: zod.string().optional(),
      texture: zod.string().optional(),
      origin: zod.string().optional(),
      isFeatured: zod.boolean(),
      inStock: zod.boolean(),
      badge: zod.string().optional(),
    }),
  ),
});

/**
 * @summary Featured / spotlight products for homepage galleries
 */
export const getFeaturedProductsQueryLimitDefault = 6;

export const GetFeaturedProductsQueryParams = zod.object({
  pole: zod
    .enum(["medical", "paramedical", "cosmetique", "supplements"])
    .optional(),
  limit: zod.coerce.number().default(getFeaturedProductsQueryLimitDefault),
});

export const GetFeaturedProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  pole: zod.enum(["medical", "paramedical", "cosmetique", "supplements"]),
  category: zod.string(),
  besoin: zod.string().optional(),
  price: zod.number(),
  imageUrl: zod.string().optional(),
  ingredients: zod.string().optional(),
  posologie: zod.string().optional(),
  contreIndications: zod.string().optional(),
  texture: zod.string().optional(),
  origin: zod.string().optional(),
  isFeatured: zod.boolean(),
  inStock: zod.boolean(),
  badge: zod.string().optional(),
});
export const GetFeaturedProductsResponse = zod.array(
  GetFeaturedProductsResponseItem,
);

/**
 * @summary List all besoin filters with product counts
 */
export const ListBesoinsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  pole: zod.string().optional(),
  productCount: zod.number(),
  icon: zod.string().optional(),
});
export const ListBesoinsResponse = zod.array(ListBesoinsResponseItem);

/**
 * @summary Get current cart contents
 */
export const GetCartQueryParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetCartResponse = zod.object({
  sessionId: zod.string(),
  items: zod.array(
    zod.object({
      id: zod.number(),
      productId: zod.number(),
      productName: zod.string(),
      productImageUrl: zod.string().optional(),
      pole: zod.string(),
      price: zod.number(),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
  itemCount: zod.number(),
});

/**
 * @summary Add item to cart
 */
export const AddToCartBody = zod.object({
  sessionId: zod.string(),
  productId: zod.number(),
  quantity: zod.number(),
});

export const AddToCartResponse = zod.object({
  sessionId: zod.string(),
  items: zod.array(
    zod.object({
      id: zod.number(),
      productId: zod.number(),
      productName: zod.string(),
      productImageUrl: zod.string().optional(),
      pole: zod.string(),
      price: zod.number(),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
  itemCount: zod.number(),
});

/**
 * @summary Remove item from cart
 */
export const RemoveFromCartParams = zod.object({
  itemId: zod.coerce.number(),
});

export const RemoveFromCartResponse = zod.object({
  sessionId: zod.string(),
  items: zod.array(
    zod.object({
      id: zod.number(),
      productId: zod.number(),
      productName: zod.string(),
      productImageUrl: zod.string().optional(),
      pole: zod.string(),
      price: zod.number(),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
  itemCount: zod.number(),
});

/**
 * @summary Upload an ordonnance (prescription)
 */
export const UploadPrescriptionBody = zod.object({
  patientName: zod.string(),
  phone: zod.string(),
  fileDataUrl: zod.string(),
  notes: zod.string().optional(),
});